+++ src/Controller/Admin/UsersController.php
@@ -23,7 +23,6 @@ use Thumber\Cake\Utility\ThumbManager;
 
 /**
  * Users controller
- * @property \MeCms\Model\Table\UsersGroupsTable $Groups
  * @property \MeCms\Controller\Component\LoginRecorderComponent $LoginRecorder
  * @property \MeTools\Controller\Component\UploaderComponent $Uploader
  * @property \MeCms\Model\Table\UsersTable $Users
@@ -46,7 +45,7 @@ class UsersController extends AppController
         }
 
         if ($this->getRequest()->is('action', ['index', 'add', 'edit'])) {
-            $groups = $this->Groups->getList()->all();
+            $groups = $this->Users->Groups->getList()->all();
             if ($groups->isEmpty()) {
                 $this->Flash->alert(__d('me_cms', 'You must first create an user group'));