+++ src/Controller/AppController.php
@@ -28,30 +28,6 @@ use RuntimeException;
 abstract class AppController extends BaseAppController
 {
     /**
-     * Magic accessor for model autoload.
-     *
-     * In addition to the method provided by CakePHP, it can also autoload the
-     *  associated tables.
-     * @param string $name Property name
-     * @return \Cake\Datasource\RepositoryInterface|\Cake\ORM\Association|null Instance or null
-     * @see \Cake\Controller\Controller::__get()
-     * @since 2.27.1
-     */
-    public function __get(string $name)
-    {
-        [, $class] = pluginSplit($this->defaultTable, true);
-
-        if ($class !== $name && $this->{$class}->hasAssociation($name)) {
-            /** @var \MeCms\Model\Table\AppTable $Table */
-            $Table = $this->{$class};
-
-            return $Table->getAssociation($name);
-        }
-
-        return parent::__get($name);
-    }
-
-    /**
      * Called before the controller action
      * @param \Cake\Event\EventInterface $event EventInterface
      * @return \Cake\Http\Response|null|void