+++ src/Controller/PagesCategoriesController.php
@@ -20,7 +20,6 @@ use MeCms\ORM\Query;
 /**
  * PagesCategories controller
  * @property \MeCms\Model\Table\PagesCategoriesTable $PagesCategories
- * @property \MeCms\Model\Table\PagesTable $Pages
  */
 class PagesCategoriesController extends AppController
 {
@@ -53,7 +52,7 @@ class PagesCategoriesController extends AppController
 
         $category = $this->PagesCategories->findActiveBySlug($slug)
             ->select(['id', 'title'])
-            ->contain($this->Pages->getAlias(), fn(Query $query): Query => $query->find('active')->select(['category_id', 'slug', 'title']))
+            ->contain($this->PagesCategories->Pages->getAlias(), fn(Query $query): Query => $query->find('active')->select(['category_id', 'slug', 'title']))
             ->cache('category_' . md5($slug))
             ->firstOrFail();